Enhance BackendMonitor: Reusability, Web3 Connectivity
Context: Backend monitor
Core: Subject: Restructure BackendMonitor, implement web3 connectivity, enhance DevOps workflows

- Refactored BackendMonitor to enhance reusability, focusing on improving readers and reporters modularity.
- Implemented logic to establish connections with web3 blockchain chains.
- Added functionality to retrieve balance information for each address from the connected web3 chains.
- Completed Docker Compose configuration to streamline local testing and development.
- Created Kubernetes .yaml files to facilitate deployment and scaling in a Kubernetes cluster.

+++ backendMonitor/src/readers/csv/CsvAccountReader.ts
@@ -3,6 +3,6 @@ import { CsvFileReader } from './CsvFileReader';
 
 export class CsvAccountReader extends CsvFileReader<Account> {
   mapRow(row: string[]): Account {
-    return new Account(row[0], parseInt(row[1]));
+    return new Account(row[0], parseFloat(row[1]));
   }
 }
